refactor(web): isolate wish estimate formatting
